packages/nfl/pyproject.toml
[project]
name = "nfl"
version = "0.1.0"
description = "NFL-specific data ingestion and feature transforms"
requires-python = ">=3.11"
dependencies = [
    "schemas",
    "transforms",
    "backends",
    "nfl-data-py>=0.3.0",
    "ibis-framework[duckdb]>=9.0.0",
    "pyarrow>=14.0.0",
    "pandas>=2.0.0",
]

[tool.uv.sources]
schemas = { workspace = true }
transforms = { workspace = true }
backends = { workspace = true }
